moved PlayerCharacter.java into new player package and refactoed BattleSimulatorActivity to MainActivity due to requirements change
