+++ app/src/main/res/layout/activity_battle_simulator.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.philipgreen.dmwizard.BattleSimulator">
+    tools:context="com.philipgreen.dmwizard.MainActivity">
 
     <Button
         android:id="@+id/character_create_button"
